requirements.txt
numpy
# pypandoc==1.4 TODO used in setup.py, requires different urlib3 than tf
seaborn==0.9.0
matplotlib==3.0.0
sklearn
kipoi>=0.6.9
kipoiseq
# torch idk if we really need this
numba # for closing gpu resources
tensorflow-gpu==1.12.0
h5sparse==0.0.5
